+++ lib/render/render.c
@@ -93,6 +93,7 @@ void renderGame(struct Game *game) {
   }
 
   if (game->state == GAME_STATE_IN_MENU) {
+    game->menu.render(game);
   }
 
   refresh();